Added movies.

+++ src/app/book-detail/book-detail.component.ts
@@ -14,8 +14,12 @@ import { CoverService } from '../cover.service';
 export class BookDetailComponent implements OnInit {
 
   @Input() book: Book;
-  origBtnText = '';
   coverUrl = '';
+  coverLoading = true;
+  defaultCoverUrl = '/assets/no-image.jpg';
+  deleteBtnText = 'Delete Book';
+  saveBtnText = 'Save Changes';
+  bookLoading = true;
 
   constructor(
     private route: ActivatedRoute,
@@ -32,44 +36,60 @@ export class BookDetailComponent implements OnInit {
   getBook(): void {
     const id = +this.route.snapshot.paramMap.get('id');
     this.bookService.getBook(id)
-      .subscribe(book =>
-        {
+      .subscribe(book => {
           this.book = book[0];
-          this.coverService.getCover(this.book.isbn_13)
-            .subscribe(res =>
-            {
-              console.log(res);
-              let items = res['items'];
-              if(typeof items !== 'undefined' && items.length > 0){
-                this.coverUrl = res['items'][0].volumeInfo.imageLinks.thumbnail;
-              }
-              else {
-                this.coverUrl = '';
-              }
-            });
-        }
-      );
+          this.bookLoading = false;
+          if (this.book.isbn_13) {
+              this.coverService.getCover(this.book.isbn_13)
+                  .subscribe(res => {
+                      const items = res['items'];
+                      if (typeof items !== 'undefined' && items.length > 0) {
+                          this.coverUrl = res['items'][0].volumeInfo.imageLinks.thumbnail;
+                      } else {
+                          this.coverUrl = this.defaultCoverUrl;
+                      }
+                      this.coverLoading = false;
+                  });
+          } else {
+              this.coverUrl = this.defaultCoverUrl;
+              this.coverLoading = false;
+          }
+      });
 
   }
 
   saveChanges(id: number): void {
+      this.saveBtnText = 'Saving Changes...';
      this.bookService.editBook(id, this.book).subscribe(data => {
      }, data => {
-       console.log('error');
+         console.log('error');
+         this.saveBtnText = 'Error Saving Changes!';
+         // 10 seconds delay
+         setTimeout( () => {
+             this.saveBtnText = 'Save Changes';
+         }, 2000 );
      }, () => {
-       console.log('edited');
+         console.log('edited');
+         this.saveBtnText = 'Saved!';
+         setTimeout(() => {
+             this.saveBtnText = 'Save Changes';
+         }, 2000);
      });
   }
 
   removeBook(id: number): void {
-      this.bookService.removeBook(id).subscribe(data => {
-
-      }, data => {
-        console.log('error');
-      }, () => {
-        console.log('deleted');
-        this.goBack();
-      });
+      if (confirm('ALERT: Are you sure to delete this book?')) {
+          this.deleteBtnText = 'Deleting..';
+          this.bookService.removeBook(id).subscribe(data => {
+          }, data => {
+              console.log('error');
+              this.deleteBtnText = 'Delete Book';
+          }, () => {
+              console.log('deleted');
+              this.deleteBtnText = 'Delete Book';
+              this.goBack();
+          });
+      }
   }
 
   goBack(): void {
